Fixed all warnings

+++ src/EntityRepresentation/Transformation.h
@@ -18,7 +18,7 @@ namespace spaceinvaders::view {
 
         Coordinate transform(spaceinvaders::Coordinate coordinate);
 
-        void transform(float &width, float &height);
+        void transform(double &width, double &height);
 
         std::shared_ptr<Dimensions> gameDimensions;
         std::shared_ptr<Dimensions> viewDimensions;
